Tidy up files a bit
